requirements.txt
fastai==2.6.3
fastcore==1.4.3
ipython==7.33.0
matplotlib==3.5.2
notebook==6.4.11
numba==0.55.1
numpy==1.21.6
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.4.1
sktime==0.11.4

# tensorflow-gpu==2.2.0
tensorflow-gpu==1.15.0
torch==1.10.1
